backend/value_creator/requirements.txt
openai>=1.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
pdf2image>=1.16.0
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
pyyaml>=6.0
requests>=2.28.0
python-dotenv>=0.19.0
tkinter-tooltip>=2.0.0
pandas>=1.3.0
reportlab>=3.6.0